requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==3.1.7
django-environ==0.9.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-gis==0.17
django-admin-list-filter-dropdown ==1.0.3
django-admin-rangefilter==0.8.4
environ==1.0
flake8==5.0.4
# GDAL @ file:///C:/Users/sergi/Desktop/django-urban-api/urban-src/GDAL-3.4.3-cp38-cp38-win_amd64.whl
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.3.0
